requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pytest
flake8
black
isort
build